refactor: consolidate date_utils w/ github
