--- packages/date_utils/package.json
@@ -1,6 +1,0 @@
-{
-	"name": "date_utils",
-	"private": true,
-	"type": "module",
-	"devDependencies": {}
-}
